+++ tests/cli/utils.py
@@ -1,6 +1,8 @@
 import asyncio
 import os
 import subprocess
+import sys
+from os.path import dirname
 
 from tests.settings import TEST_DATABASE
 
@@ -9,12 +11,12 @@ def run_cmd(app, cmd, with_app_environment=True, extra_env=None):
     env = dict(os.environ)
     # for main.py
     env["TEST_DATABASE"] = TEST_DATABASE
+    env["PATH"] = f"{dirname(sys.executable)}:{env['PATH']}"
     env.setdefault("PYTHONPATH", env["PWD"])
     if with_app_environment:
         env["EDGY_DEFAULT_APP"] = app
     if extra_env:
         env.update(extra_env)
-    cmd = f"hatch --env test run {cmd}"
 
     result = subprocess.run(cmd, capture_output=True, env=env, shell=True)
     print("\n$ " + cmd)
@@ -27,12 +29,12 @@ async def arun_cmd(app, cmd, with_app_environment=True, extra_env=None):
     env = dict(os.environ)
     # for main.py
     env["TEST_DATABASE"] = TEST_DATABASE
+    env["PATH"] = f"{dirname(sys.executable)}:{env['PATH']}"
     env.setdefault("PYTHONPATH", env["PWD"])
     if with_app_environment:
         env["EDGY_DEFAULT_APP"] = app
     if extra_env:
         env.update(extra_env)
-    cmd = f"hatch --env test run {cmd}"
 
     process = await asyncio.create_subprocess_shell(
         cmd, env=env, stderr=subprocess.PIPE, stdout=subprocess.PIPE